requirements.txt
fastapi==0.110.0
uvicorn>=0.32.0
SQLAlchemy>=2.0.40
aiosqlite==0.19.0
pydantic>=2.8
pydantic-settings>=2.0
jinja2>=3.1.6
APScheduler==3.10.4
openpyxl>=3.1.2
imapclient==2.3.1
requests>=2.31.0
PyYAML>=6.0.1
python-multipart==0.0.9
email-validator==2.1.1
beautifulsoup4>=4.12
